more form cleanups more uses of fidelity validator moved validators to own module added more uses of read_form_field
